Combine all mock modules into one
