--- tests/mock_cell_parser.py
@@ -1,7 +1,0 @@
-class MockCellParser:
-
-    def parse(self, value, context={}):
-        return value
-
-    def parse_as_string(self, value, context={}):
-        return value
