+++ tests/test_rowdatasheet.py
@@ -5,8 +5,8 @@ from pathlib import Path
 from tempfile import TemporaryDirectory
 from typing import List, Dict, Optional
 
-from tests.mock_row_parser import MockRowParser
 from rpft.parsers.common.rowdatasheet import RowDataSheet
+from tests.mocks import MockRowParser
 
 
 rowA = OrderedDict([
